# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for NutriPlanner - Nutritionist and Patient Platform
 * OpenAPI spec version: 0.1.0
 */
import type { PatientGender } from "./patientGender";

export interface Patient {
  id: number;
  name: string;
  email: string;
  phone?: string;
  birthDate?: Date;
  gender?: PatientGender;
  /** Height in cm */
  height?: number;
  /** Weight in kg */
  weight?: number;
  objective?: string;
  observations?: string;
  nutritionistId: number;
  createdAt: Date;
}